Added basic logging

+++ src/com/mt/worldgen/pipeline/ConsoleProgressMonitor.java
@@ -5,7 +5,7 @@ public class ConsoleProgressMonitor implements ProgressListener {
   @Override
   public void onProgressUpdate(int currentStageIndex, int totalStages, String stageName, String status) {
     double percentage = (double) currentStageIndex / totalStages * 100;
-    String statusMessage = MessageFormat.format("{0}% step {1}/{2}: {3}",percentage,currentStageIndex,totalStages,status);
+    String statusMessage = MessageFormat.format("{0}% step {1}/{2}: {3}",(int)percentage,currentStageIndex,totalStages,status);
     System.out.println(statusMessage);
   }
 }
